Fix cross-attn norm3 output dtype: cast to float16 for quant_sum
nn.LayerNorm outputs float32 by default, but fused_kernels.quant_sum
requires float16 input. Add explicit .to(torch.float16) after norm3.

+++ examples/wan2.2_fun/models/quant_wan_cuda.py
@@ -306,7 +306,7 @@ class WanAttentionBlockWithCudaKernel(nn.Module):
 
         # ===== Cross-Attention =====
         residual = x
-        x_norm = self.norm3(x)
+        x_norm = self.norm3(x).to(torch.float16)
         # Quantize image features for cross-attention Q
         x_quant = fused_kernels.quant_sum(
             x_norm.contiguous(),
